Refactor: courses.session 패키지 추가
