+++ src/main/java/nextstep/courses/domain/session/Session.java
@@ -1,9 +1,8 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import nextstep.courses.domain.enrollment.SessionPeriod;
 import nextstep.courses.domain.enrollment.engine.SessionEnrollment;
 import nextstep.courses.domain.image.SessionCoverImage;
-import nextstep.courses.domain.status.ProgressStatus;
 import nextstep.courses.domain.student.SessionStudent;
 import nextstep.courses.infrastructure.entity.BaseEntity;
 import nextstep.payments.domain.Payment;
