+++ src/main/java/nextstep/courses/infrastructure/engine/SessionRepository.java
@@ -1,6 +1,6 @@
 package nextstep.courses.infrastructure.engine;
 
-import nextstep.courses.domain.Session;
+import nextstep.courses.domain.session.Session;
 
 import java.util.Optional;
