+++ src/test/java/nextstep/courses/domain/status/SessionProgressStatusTest.java
@@ -1,5 +1,6 @@
 package nextstep.courses.domain.status;
 
+import nextstep.courses.domain.session.ProgressStatus;
 import nextstep.courses.exception.SessionProgressStatusInvalidException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -9,7 +10,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
 
-import static nextstep.courses.domain.status.ProgressStatus.*;
+import static nextstep.courses.domain.session.ProgressStatus.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
