+++ src/test/java/nextstep/courses/service/SessionServiceTest.java
@@ -1,7 +1,7 @@
 package nextstep.courses.service;
 
-import nextstep.courses.domain.Session;
-import nextstep.courses.domain.SessionType;
+import nextstep.courses.domain.session.Session;
+import nextstep.courses.domain.session.SessionType;
 import nextstep.courses.domain.fixture.SessionFixture;
 import nextstep.courses.domain.student.SessionStudents;
 import nextstep.courses.exception.SessionNotFoundException;
